Rework state into event handler
The previous implementation was difficult to prototype with because I
had to reload the server and client modules in order to reset the atoms
into a initial state.

This diff reworks that logic into the server and client `handler`
functions. The approach works also scales well because `atoms` should be
able to sync multiple clients by adding additional watches.