Fixed bug where repeated versions are accepted
This re-works the implementation so that repeated versions are
ignored. Previously, the local version of the entangle was not updated
if the patch was empty. An empty patch can be used to signify that the
synchronization has reached a stead state.

Unfortunately, the local version did not always increment when data was
received, which allowed two of the same patches to be applied.

Some small minor tweaks to comments as well.